feat: project is now a thin page with sections
