+++ admin/src/__tests__/integration/ProjectReleasesSection.test.tsx
@@ -1,0 +1,204 @@
+import { fireEvent, screen, waitFor } from '@solidjs/testing-library';
+import { http, HttpResponse } from 'msw';
+import { beforeEach, describe, expect, it } from 'vitest';
+
+import { server } from '../mocks/server';
+import {
+  renderProjectSections,
+  resetProjectSectionsTestState,
+} from './project-sections.test-utils';
+
+describe('ProjectReleasesSection', () => {
+  beforeEach(() => {
+    resetProjectSectionsTestState();
+  });
+
+  it('publishes a configuration release without auto-activating when a release is already active', async () => {
+    const publishRequests: Array<{ version: string; makeActive: boolean }> = [];
+    server.use(
+      http.post(
+        'http://localhost:5027/admin/projects/:projectId/environments/:envName/releases',
+        async ({ params, request }) => {
+          const body = (await request.json()) as { version: string; makeActive: boolean };
+          publishRequests.push(body);
+          return HttpResponse.json(
+            {
+              project: params.projectId,
+              environment: params.envName,
+              version: body.version,
+              entryCount: 3,
+              isActive: body.makeActive,
+              createdAt: new Date().toISOString(),
+              actor: 'admin@example.com',
+              entries: [],
+            },
+            { status: 201 },
+          );
+        },
+      ),
+    );
+
+    renderProjectSections('/projects/my-app/releases');
+
+    fireEvent.click(await screen.findByTestId('release-create-version-button'));
+    fireEvent.input(await screen.findByTestId('release-version-input'), {
+      target: { value: '1.2' },
+    });
+    fireEvent.click(screen.getByTestId('release-version-confirm-button'));
+
+    fireEvent.click(await screen.findByTestId('release-create-confirm-button'));
+
+    await waitFor(() => {
+      expect(publishRequests).toEqual([{ version: '1.2.0', makeActive: false }]);
+    });
+  });
+
+  it('keeps you on the parameters step when creating the release fails', async () => {
+    server.use(
+      http.post(
+        'http://localhost:5027/admin/projects/:projectId/environments/:envName/releases',
+        () => HttpResponse.json({ detail: 'Release already exists' }, { status: 409 }),
+      ),
+    );
+
+    renderProjectSections('/projects/my-app/releases');
+
+    fireEvent.click(await screen.findByTestId('release-create-version-button'));
+    fireEvent.input(await screen.findByTestId('release-version-input'), {
+      target: { value: '1.2' },
+    });
+    fireEvent.click(screen.getByTestId('release-version-confirm-button'));
+
+    fireEvent.click(await screen.findByTestId('release-create-confirm-button'));
+
+    expect(await screen.findByText('Release already exists')).toBeInTheDocument();
+    expect(screen.getByTestId('release-create-confirm-button')).toBeInTheDocument();
+  });
+
+  it('amends a release into a new patch without touching working config', async () => {
+    const draftCalls: string[] = [];
+    const publishRequests: Array<{
+      version: string;
+      makeActive: boolean;
+      entries?: Array<{ key: string; value: string }>;
+    }> = [];
+    server.use(
+      http.post(
+        'http://localhost:5027/admin/projects/:projectId/environments/:envName/releases/:version/draft',
+        ({ params }) => {
+          draftCalls.push(String(params.version));
+          return HttpResponse.json([]);
+        },
+      ),
+      http.get(
+        'http://localhost:5027/admin/projects/:projectId/environments/:envName/releases/:version',
+        ({ params }) =>
+          HttpResponse.json({
+            project: params.projectId,
+            environment: params.envName,
+            version: params.version,
+            entryCount: 1,
+            isActive: false,
+            createdAt: '2024-01-01T00:00:00Z',
+            actor: 'alice',
+            entries: [
+              { key: 'feature.x', value: 'true', contentType: 'boolean', scope: 'client' },
+            ],
+          }),
+      ),
+      http.post(
+        'http://localhost:5027/admin/projects/:projectId/environments/:envName/releases',
+        async ({ request }) => {
+          publishRequests.push((await request.json()) as (typeof publishRequests)[number]);
+          return HttpResponse.json({}, { status: 201 });
+        },
+      ),
+    );
+
+    renderProjectSections('/projects/my-app/releases');
+
+    fireEvent.click(await screen.findByTestId('release-amend-1.1.0'));
+    expect(screen.queryByTestId('release-version-dialog')).not.toBeInTheDocument();
+
+    await screen.findByTestId('amend-row-feature.x');
+
+    fireEvent.click(screen.getByTestId('release-amend-confirm-button'));
+    await waitFor(() => {
+      expect(publishRequests.length).toBe(1);
+    });
+    expect(publishRequests[0].version).toBe('1.1.1');
+    expect(publishRequests[0].makeActive).toBe(false);
+    expect(publishRequests[0].entries?.some(entry => entry.key === 'feature.x')).toBe(true);
+    expect(draftCalls).toEqual([]);
+  });
+
+  it('activates a configuration release', async () => {
+    const activeRequests: Array<{ version: string | null }> = [];
+    server.use(
+      http.put(
+        'http://localhost:5027/admin/projects/:projectId/environments/:envName/active-release',
+        async ({ params, request }) => {
+          const body = (await request.json()) as { version: string | null };
+          activeRequests.push(body);
+          return HttpResponse.json({
+            project: params.projectId,
+            name: params.envName,
+            activeReleaseVersion: body.version,
+            createdAt: '2024-01-01T00:00:00Z',
+            updatedAt: new Date().toISOString(),
+          });
+        },
+      ),
+    );
+
+    renderProjectSections('/projects/my-app/releases');
+
+    await screen.findByText('1.1.0');
+    const activateButtons = screen.getAllByRole('button', { name: /activate/i });
+    const enabledActivate = activateButtons.find(button => !button.hasAttribute('disabled'));
+    expect(enabledActivate).toBeTruthy();
+    fireEvent.click(enabledActivate!);
+    expect(await screen.findByTestId('release-activate-dialog')).toBeInTheDocument();
+    fireEvent.click(screen.getByTestId('release-activate-confirm-button'));
+
+    await waitFor(() => {
+      expect(activeRequests).toEqual([{ version: '1.1.0' }]);
+    });
+  });
+
+  it('opens release parameters on the parameters page without starting an amend flow', async () => {
+    renderProjectSections('/projects/my-app/releases');
+
+    fireEvent.click(await screen.findByTestId('release-view-1.1.0'));
+
+    expect(await screen.findByTestId('release-view-banner')).toBeInTheDocument();
+    expect(await screen.findByTestId('project-parameters-heading')).toBeInTheDocument();
+    expect(await screen.findByTestId('parameter-row-API_URL')).toBeInTheDocument();
+    expect(screen.queryByTestId('release-create-confirm-button')).not.toBeInTheDocument();
+  });
+
+  it('deletes a non-active release after confirmation', async () => {
+    const deleteRequests: string[] = [];
+    server.use(
+      http.delete(
+        'http://localhost:5027/admin/projects/:projectId/environments/:envName/releases/:version',
+        ({ params }) => {
+          deleteRequests.push(String(params.version));
+          return new HttpResponse(null, { status: 204 });
+        },
+      ),
+    );
+
+    renderProjectSections('/projects/my-app/releases');
+
+    const activeDeleteButton = await screen.findByTestId('release-delete-1.0.0');
+    expect(activeDeleteButton).toBeDisabled();
+    fireEvent.click(screen.getByTestId('release-delete-1.1.0'));
+    expect(await screen.findByTestId('release-delete-dialog')).toBeInTheDocument();
+    fireEvent.click(screen.getByTestId('release-delete-confirm-button'));
+
+    await waitFor(() => {
+      expect(deleteRequests).toEqual(['1.1.0']);
+    });
+  });
+});
